backend/requirements.txt
# --- Core Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# --- Database & ORM ---
SQLAlchemy>=2.0.29
# Tùy chọn 1: Driver cho MySQL (PyMySQL là pure python, dễ đóng gói bằng Tauri/PyInstaller)
pymysql>=1.1.0 
# Tùy chọn 2: Driver cho SQL Server (yêu cầu cài đặt ODBC driver trên máy)
# pyodbc>=5.1.0 

# --- AI & LLM (LangChain Ecosystem) ---
langchain>=0.1.13
langchain-groq>=0.1.0
langchain-community>=0.0.29

# --- Utilities ---
python-dotenv>=1.0.1
cryptography>=42.0.0
pyinstaller>=6.5.0